[FEATURE] chosen library to application. Added tags and branches to deploy view.
This patch adds the API connection to retrieve tags and branches on a
deployment view of a project from gitlab.

It also introduced the chosen library for advanced select boxes. To use
chosen, jQuery is requiered. Both libraries are added to index.html.

A chosaen directive is also added to the angularJS application to trigger
chosen after asynchronous ajax calls.

Versio raised to 0.4.0